Major: Fixed optimizer calls AND major typo conidtional
First, I had omitted 'not' in a conditional, which broke most everything. Added 'not' back into 'if self.rebalance is not None'.
Second, there was a construct issue in how I added the case-switch dict for the optimizers. This should now be resolved in both locations.
Third, minor: renamed 'information_ratio()' to 'calculate_information_ration()', as the internal variables shadowed the function name. Using the guide that variables should be nouns, and functions should be verbs.